refactor(rendering): move CameraSnapshot to svg (common = both engines)
Made-with: Cursor
