+++ src/rendering/common/camera/Camera.ts
@@ -142,6 +142,3 @@ export class Camera {
     this.rotation = this.defaultRotation;
   }
 }
-
-/** Serializable view state (`x`/`y` pan, `zoom`, `rotation`); use with {@link Camera} via `readTransform`-style copy — not a class instance. */
-export type CameraSnapshot = Pick<Camera, 'x' | 'y' | 'zoom' | 'rotation'>;
