+++ src/rendering/svg/components/SvgBoard.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import { Board } from '../../../models/Board';
 import { radiansToDegrees } from '../../../utils/Angle';
-import type { CameraSnapshot } from '../../common/camera/Camera';
+import type { CameraSnapshot } from '../cameraSnapshot';
 import { HexTile } from '../tiles/HexTile';
 import { SVG_HEX_RADIUS, hexToPixel } from '../tiles/SvgHexUtils';
